Summary: Updated acceptance criteria
Description: Updated the Address model to include a list of available states, and the Deal model to check for updated acceptance criteria. I chose to restrict the user to only the dates, state(s), and money value that matched the criteria when they submit their form. It didn't make sense to allow the user to have all the states to choose from but only 1 valid state. Same concept with the date range, and adding a 'max' value to the number field in the form gives them immediate form validation without submitting and reloading the page.